Export just top-level files and directories from directories
The `EXPORT_CONTENT` variables makes sure that no symlinks are made to
directories at all under the directories specified - rather, all
directories are created in the destination, and only files are
symlinked. Put the contents of the directories in this repository in the
`EXPORT_APPEND` using wildcards, so that files and directories contained
directly under the top-level directories are symlinked.