Fix org-git-push to work correctly with multiple push-urls in remote
`git pull` wasn't updating the status of the remote tracking branches
correctly, so `git push --force-with-lease` refused to push to them in some
cases, because it believed it would overwrite remote commits, despite the fact
that I was pushing to all the affected remotes. Work around this by manually
tagging the HEAD of the current branch and telling `--force-with-lease` to
ensure that the remote branches include the current commit before pushing
them.