open-in-editor: fix shebang
On macOS (and many other systems), an executable in the shebang line
will only be passed a maximum of one argument. Therefore, in
`#!/usr/bin/env bash -l`, the `-l` will be ignored, and so `PATH` won't
contain the correct binaries. Fix this by removing `-l` and setting
`PATH` explicitly.