Add Iterm2 Shell Integration
While we were loading it before if installed locally, it turns out that
it was breaking bash-preexec because the version installed in iTerm used
a difference include guard from upstream. To avoid this in the future,
always load bash-preexec from the iTerm2 shell extensions, which can
also be installed on Linux machines.

We first extract bash-preexec from the iTerm2 shell extensions and
install just bash-preexec. That way, if the iTerm2 shell extensions fail
to load for some reason, we still have bash-preexec loaded.