Git: disable core.excludesFile, move `ignore` to repo template
The fact that the ignore file is pretty wide-ranging has caused problems
in repositories that I've cloned from others, since files that might
influence the builds in those repositories and shouldn't be ignored are
being ignored. Instead, convert it to a .gitignore in the template used
when starting a new repository.